+++ client/src/config/eventSearchTags.ts
@@ -54,6 +54,11 @@ export const relatedAccordions: IHalLinks = {
     searchTag: 'lux:eventConceptsInfluencedBy',
     tab: 'concepts',
   },
+  agents: {
+    title: 'Related People and Groups',
+    searchTag: 'lux:eventRelatedAgents',
+    tab: 'people',
+  },
   materials: {
     title: 'Related Materials',
     searchTag: 'lux:eventItemMaterials',